erofs-utils: complete extended inode support
There is no requirement to use extended inode (v2)
for Android, but it's obviously useful for wider
use cases.

extended inode was partially supported by obsoleted_mkfs,
complete for new erofs-utils as well.

Tested-by: Li Guifu <blucerlee@gmail.com>
Link: https://lore.kernel.org/r/20191004204630.22696-2-hsiangkao@aol.com

+++ lib/compress.c
@@ -423,8 +423,8 @@ int erofs_write_compressed_file(struct erofs_inode *inode)
 	remaining = inode->i_size;
 
 	while (remaining) {
-		const uint readcount = min_t(uint, remaining,
-					     sizeof(ctx.queue) - ctx.tail);
+		const u64 readcount = min_t(u64, remaining,
+					    sizeof(ctx.queue) - ctx.tail);
 
 		ret = read(fd, ctx.queue + ctx.tail, readcount);
 		if (ret != readcount) {
